Templates überarbeitet auf neue Umfragenstruktur. Generierung der DTOs in ApplicationService gezogen. Diverse Kleinigkeiten für die ich keinen Bock hab nen eigenen Commit zu erstellen.

+++ src/main/java/de/wps/bikehh/umfragenerstellen/controller/UmfrageErstellenController.java
@@ -75,7 +75,7 @@ public class UmfrageErstellenController {
 	 * @return HTML-Template
 	 */
 	@GetMapping("/bearbeiten")
-	public String zeigeUmfragenBearbeiter(@RequestParam(required = true, name = "umfrageId") int umfrageId,
+	public String zeigeUmfragenBearbeiter(@RequestParam(required = true, name = "umfrageId") Long umfrageId,
 			Model model) {
 
 		Umfrage umfrage = umfragenService.getById(umfrageId);
